move the common code to a separated file
