+++ erpc_c/setup/erpc_server_setup.cpp
@@ -33,6 +33,7 @@
 #include "basic_codec.h"
 #include "message_buffer.h"
 #include "erpc_config_internal.h"
+#include "erpc_setup.h"
 #include <new>
 #include <assert.h>
 
@@ -46,24 +47,6 @@ using namespace erpc;
 // Classes
 ////////////////////////////////////////////////////////////////////////////////
 
-class BasicMessageBufferFactory : public MessageBufferFactory
-{
-public:
-    virtual MessageBuffer create()
-    {
-        uint8_t *buf = new (nothrow) uint8_t[ERPC_DEFAULT_BUFFER_SIZE];
-        return MessageBuffer(buf, ERPC_DEFAULT_BUFFER_SIZE);
-    }
-
-    virtual void dispose(MessageBuffer *buf)
-    {
-        assert(buf);
-        if (*buf)
-        {
-            delete[] buf->get();
-        }
-    }
-};
 
 ////////////////////////////////////////////////////////////////////////////////
 // Variables
